feat: upgrade Gateway API to v1.6.0 and policy-machinery to v0.10.0
- Upgrade sigs.k8s.io/gateway-api to v1.6.0 and github.com/kuadrant/policy-machinery to v0.10.0
- Promote TargetRef fields across core & extension policies to gatewayapiv1.LocalPolicyTargetReferenceWithSectionName
- Promote policy condition/reason constants to gatewayapiv1
- Update listener TLS configuration references to gatewayapiv1.ListenerTLSConfig
- Regenerate CRDs, manifests, bundle, and helm charts

+++ internal/extension/reconciler.go
@@ -271,7 +271,7 @@ func toListeners(listeners []v1.Listener) []*extpb.Listener {
 		if l.Protocol != "" {
 			listener.Protocol = string(l.Protocol)
 		}
-		listener.Port = int32(l.Port)
+		listener.Port = l.Port
 		ls[i] = &listener
 	}
 	return ls
